requirements.txt
certifi==2019.9.11
chardet==3.0.4
gitdb2==2.0.6
GitPython==3.0.4
idna==2.8
mistune==0.8.4
PyYAML==5.1.2
requests==2.22.0
smmap2==2.0.5
textwrap3==0.9.2
urllib3==1.25.7
yapf==0.28.0
